# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Inkwell API
  version: "0.2"
  description: |
    API-first Markdown publishing platform.

    All write endpoints require `X-Api-Key`. Reads are public unless stated
    otherwise. See [docs/API.md](./API.md) for the full narrative reference.
  contact:
    name: Inkwell
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-railway-domain.up.railway.app
    description: Production (replace with actual domain)
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Shared admin key (`INKWELL_API_KEY`) or a scoped token
        (`ink_<prefix>_<secret>`).
  schemas:
...
    AskRequest:
      type: object
      required: [q]
      properties:
        q:
          type: string
          maxLength: 1000
          example: "What is a digital garden?"
    AskResponse:
      type: object
      required: [query, answer, citations]
      properties:
        query:
          type: string
        answer:
          type: string
        citations:
          type: array
          items:
            type: object
            required: [slug, title]
            properties:
              slug:
                type: string
              title:
                type: string
    RelatedResponse:
      type: object
      required: [slug, related]
      properties:
        slug:
          type: string
        related:
          type: array
          items:
            type: object
            required: [slug, title, distance]
            properties:
              slug:
                type: string
              title:
                type: string
              distance:
                type: number
                format: double
                description: Cosine distance (lower = more similar).
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
              example: "Missing or invalid API key."
            requestId:
              type: string
              format: uuid
              description: Correlation ID. Present on every in-request error.
            slug:
              type: string
              description: Present on 409 Conflict for duplicate slug.
            allow:
              type: string
              description: Present on 405 Method Not Allowed.
              example: "GET, POST"
  responses:
...
    RejectedReadCredential:
      description: >
        An `X-Api-Key` was presented and rejected (revoked token, unknown prefix,
        hash mismatch, a value that is neither the shared key nor token-shaped, or
        a duplicated/non-ASCII header). Treat this as "re-authenticate", not
        "empty result" — see ADR 0015. A rejected `inkwell_session` cookie is NOT
        reported here: that channel downgrades to anonymous and returns `200` with
        published content, so a stale browser cookie never breaks public pages. A
        valid credential lacking the `read` scope also stays `200`, published only.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Missing or invalid API key."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    ServiceUnavailable:
      description: >
        The credential could not be judged because the token lookup failed against
        Postgres (ADR 0015). Distinct from `401` — the credential was never
        actually rejected — and from `500`. Retry; check `GET /readyz`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Service temporarily unavailable. Retry shortly."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
...
    NotFound:
      description: Resource does not exist (or caller cannot see it).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "No document with slug \"hello-world\"."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Rate limit exceeded. Slow down and retry later."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
paths:
...
  /documents/{slug}/related:
    get:
      summary: Semantically related notes
      operationId: documentRelated
      tags: [AI]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Top 5 related notes by embedding cosine distance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatedResponse"
              example:
                slug: "hello-world"
                related:
                  - slug: "rust-notes"
                    title: "Rust Notes"
                    distance: 0.12
        "404":
          $ref: "#/components/responses/NotFound"
...
  /ask:
    get:
      summary: AI question answering (GET)
      operationId: askGet
      tags: [AI]
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            maxLength: 1000
          example: "What is a digital garden?"
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Grounded answer with citations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
              example:
                query: "What is a digital garden?"
                answer: "A digital garden is an evolving collection of interconnected notes…"
                citations:
                  - slug: "digital-garden"
                    title: "Digital Garden"
        "400":
          description: Empty or oversized query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      summary: AI question answering (POST)
      operationId: askPost
      tags: [AI]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
            example:
              q: "What is a digital garden?"
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Grounded answer with citations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
        "400":
          description: Empty or oversized query, or invalid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...